chore: refactor schedule system from donut
